requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiostream==0.4.5
alembic==1.11.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
apache-airflow==2.6.2
apache-airflow-providers-common-sql==1.5.2
apache-airflow-providers-ftp==3.4.2
apache-airflow-providers-http==4.4.2
apache-airflow-providers-imap==3.2.2
apache-airflow-providers-sqlite==3.4.2
apispec==5.2.2
appdirs==1.4.4
appnope==0.1.3
argcomplete==3.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.1
asteroid-filterbanks==0.4.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
av==10.0.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
blis==0.7.9
bokeh==2.4.3
boto3==1.26.112
botocore==1.29.112
bs4==0.0.1
cachelib==0.9.0
catalogue==2.0.8
cattrs==23.1.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-default-group-wheel==1.2.2
clickclick==20.10.2
cloudpickle==2.0.0
cmaes==0.9.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==4.8.0
comm==0.1.3
commonmark==0.9.1
confection==0.1.0
ConfigUpdater==3.1.1
#connectorx==0.3.1
connexion==2.14.2
contourpy==1.0.7
cron-descriptor==1.4.0
croniter==1.3.15
cryptography==41.0.1
ctranslate2==3.16.0
cycler==0.11.0
cymem==2.0.7
dask==2023.3.2
dateutils==0.6.12
debugpy==1.6.6
decorator==5.1.1
deepdiff==6.3.0
defusedxml==0.7.1
deltalake==0.8.1
Deprecated==1.2.14
dill==0.3.6
distributed==2023.3.2
dnspython==2.3.0
docopt==0.6.2
docutils==0.20.1
duckdb==0.7.1
einops==0.6.1
email-validator==1.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.2
fastapi-utils==0.2.1
faster-whisper==0.6.0
fastjsonschema==2.16.3
feedparser==6.0.10
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.12.2
Flask==2.2.3
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flatbuffers==23.5.26
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
graphviz==0.20.1
grpclib==0.4.3
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
HeapDict==1.0.1
hmmlearn==0.3.0
hpack==4.0.0
html5lib==1.1
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
humanfriendly==10.0
hyperframe==6.0.1
HyperPyYAML==1.2.1
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
inflection==0.5.1
inquirer==3.1.3
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
julius==0.2.7
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
langcodes==3.3.0
lazy-object-proxy==1.9.0
lazy_loader==0.3
librosa==0.10.0.post2
lightning==2.0.4
lightning-cloud==0.5.37
lightning-utilities==0.9.0
limits==3.5.0
linkify-it-py==2.0.2
llm==0.3
llvmlite==0.40.1
load-dotenv==0.1.0
locket==1.0.0
lockfile==0.12.2
lxml==4.9.2
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==2.0.5
modal-client==0.49.2289
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
nba-api==1.2
nbclassic==0.5.4
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numba==0.57.1
numpy==1.24.2
omegaconf==2.3.0
onnxruntime==1.15.1
openai==0.27.2
opencv-python-headless==4.7.0.72
optuna==3.2.0
ordered-set==4.1.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pathspec==0.9.0
pathy==0.10.2
pendulum==2.1.2
pexpect==4.8.0
pgvector==0.1.8
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
plotly==5.14.0
pluggy==1.2.0
poetry-core==1.5.2
polars==0.16.17
pooch==1.6.0
preshed==3.0.8
primePy==1.3
prison==0.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.1
psutil==5.9.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyarrow==11.0.0
pycparser==2.21
pydantic==1.10.7
pyfredapi==0.5.1
Pygments==2.14.0
PyJWT==2.7.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
python-nvd3==0.15.0
python-slugify==8.0.1
pytorch-lightning==2.0.4
pytorch-metric-learning==2.2.0
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.1
rapidfuzz==3.1.1
readchar==4.0.5
regex==2023.5.5
requests==2.28.2
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rich_argparse==1.1.1
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.9.0
semver==3.0.1
Send2Trash==1.8.0
sentencepiece==0.1.99
setproctitle==1.3.2
sgmllib3k==1.0.0
shellingham==1.5.0.post1
sigtools==4.0.1
six==1.16.0
scikit-learn
slack-bolt==1.18.0
slack-sdk==3.21.3
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4
soxr==0.3.5
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spaczz==0.6.0
speechbrain==0.5.14
SQLAlchemy==1.4.48
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlite-fts4==1.0.3
sqlite-utils==3.31
sqlparse==0.4.4
srsly==2.4.6
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
stomper==0.4.3
strip-tags==0.2
symbex==1.0
sympy==1.12
synchronicity==0.5.2
tabulate==0.9.0
tblib==1.7.0
tenacity==8.2.2
tensorboardX==2.6.1
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
thinc==8.1.10
threadpoolctl==3.1.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
trio==0.22.0
trio-websocket==0.10.2
ttok==0.1
typer==0.9.0
types-certifi==2021.10.8.3
types-toml==0.10.8.6
typing_extensions==4.5.0
uc-micro-py==1.0.2
undetected-chromedriver==3.2.1
unicodecsv==0.14.1
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.22.0
wasabi==1.1.2
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.2
Werkzeug==2.2.3
whisperx @ git+https://github.com/m-bain/whisperx.git@befe2b242eb59dcd7a8a122d127614d5c63d36e9
widgetsnbextension==4.0.7
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
xlsx2csv==0.8.1
XlsxWriter==3.0.9
yarl==1.8.2
zict==2.2.0
zipp==3.15.0
